Deprecate `:confirm` in favor of `:data => { :confirm => 'Text' }` option
This deprecation applies to:
`button_to`
`button_tag`
`image_submit_tag`
`link_to`
`submit_tag`

As :confirm is an UI specific option is better to use the data attributes,
teaching users about unobtrusive JavaScript and how Rails works with it.

+++ actionpack/test/template/url_helper_test.rb
@@ -75,9 +75,18 @@ class UrlHelperTest < ActiveSupport::TestCase
   end
 
   def test_button_to_with_javascript_confirm
+    assert_deprecated ":confirm option is deprecated and will be removed from Rails 4.0. Use ':data => { :confirm => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\"><div><input data-confirm=\"Are you sure?\" type=\"submit\" value=\"Hello\" /></div></form>",
+        button_to("Hello", "http://www.example.com", :confirm => "Are you sure?")
+      )
+    end
+  end
+
+  def test_button_to_confirm_without_deprecation_warning
     assert_dom_equal(
       "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\"><div><input data-confirm=\"Are you sure?\" type=\"submit\" value=\"Hello\" /></div></form>",
-      button_to("Hello", "http://www.example.com", :confirm => "Are you sure?")
+      button_to("Hello", "http://www.example.com", :data => { :confirm => "Are you sure?" })
     )
   end
 
@@ -95,9 +104,18 @@ class UrlHelperTest < ActiveSupport::TestCase
   end
 
   def test_button_to_with_remote_and_javascript_confirm
+    assert_deprecated ":confirm option is deprecated and will be removed from Rails 4.0. Use ':data => { :confirm => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\" data-remote=\"true\"><div><input data-confirm=\"Are you sure?\" type=\"submit\" value=\"Hello\" /></div></form>",
+        button_to("Hello", "http://www.example.com", :remote => true, :confirm => "Are you sure?")
+      )
+    end
+  end
+
+  def test_button_to_with_remote_and_javascript_confirm_without_deprecation_warning
     assert_dom_equal(
       "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\" data-remote=\"true\"><div><input data-confirm=\"Are you sure?\" type=\"submit\" value=\"Hello\" /></div></form>",
-      button_to("Hello", "http://www.example.com", :remote => true, :confirm => "Are you sure?")
+      button_to("Hello", "http://www.example.com", :remote => true, :data => { :confirm => "Are you sure?" })
     )
   end
 
@@ -110,13 +128,20 @@ class UrlHelperTest < ActiveSupport::TestCase
     end
   end
 
-  def test_button_to_with_remote_and_javascript_confirm_and_javascript_disable_with
-    assert_deprecated ":disable_with option is deprecated and will be removed from Rails 4.0. Use 'data-disable-with' instead" do
+  def test_button_to_with_remote_and_javascript_confirm
+    assert_deprecated ":confirm option is deprecated and will be removed from Rails 4.0. Use ':data => { :confirm => \'Text\' }' instead" do
       assert_dom_equal(
-        "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\" data-remote=\"true\"><div><input data-disable-with=\"Greeting...\" data-confirm=\"Are you sure?\" type=\"submit\" value=\"Hello\" /></div></form>",
-        button_to("Hello", "http://www.example.com", :remote => true, :confirm => "Are you sure?", :disable_with => "Greeting...")
+        "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\" data-remote=\"true\"><div><input data-confirm=\"Are you sure?\" type=\"submit\" value=\"Hello\" /></div></form>",
+        button_to("Hello", "http://www.example.com", :remote => true, :confirm => "Are you sure?")
       )
     end
+  end
+
+  def test_button_to_with_remote_and_javascript_confirm_without_deprecation_warning
+    assert_dom_equal(
+      "<form method=\"post\" action=\"http://www.example.com\" class=\"button_to\" data-remote=\"true\"><div><input data-confirm=\"Are you sure?\" type=\"submit\" value=\"Hello\" /></div></form>",
+      button_to("Hello", "http://www.example.com", :remote => true, :data => { :confirm => "Are you sure?" })
+    )
   end
 
   def test_button_to_with_remote_false
@@ -206,17 +231,30 @@ class UrlHelperTest < ActiveSupport::TestCase
   end
 
   def test_link_tag_with_javascript_confirm
+    assert_deprecated ":confirm option is deprecated and will be removed from Rails 4.0. Use ':data => { :confirm => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<a href=\"http://www.example.com\" data-confirm=\"Are you sure?\">Hello</a>",
+        link_to("Hello", "http://www.example.com", :confirm => "Are you sure?")
+      )
+    end
+    assert_deprecated ":confirm option is deprecated and will be removed from Rails 4.0. Use ':data => { :confirm => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<a href=\"http://www.example.com\" data-confirm=\"You can't possibly be sure, can you?\">Hello</a>",
+        link_to("Hello", "http://www.example.com", :confirm => "You can't possibly be sure, can you?")
+      )
+    end
+    assert_deprecated ":confirm option is deprecated and will be removed from Rails 4.0. Use ':data => { :confirm => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<a href=\"http://www.example.com\" data-confirm=\"You can't possibly be sure,\n can you?\">Hello</a>",
+        link_to("Hello", "http://www.example.com", :confirm => "You can't possibly be sure,\n can you?")
+      )
+    end
+  end
+
+  def test_link_tag_confirm_without_deprecation_warning
     assert_dom_equal(
       "<a href=\"http://www.example.com\" data-confirm=\"Are you sure?\">Hello</a>",
-      link_to("Hello", "http://www.example.com", :confirm => "Are you sure?")
-    )
-    assert_dom_equal(
-      "<a href=\"http://www.example.com\" data-confirm=\"You can't possibly be sure, can you?\">Hello</a>",
-      link_to("Hello", "http://www.example.com", :confirm => "You can't possibly be sure, can you?")
-    )
-    assert_dom_equal(
-      "<a href=\"http://www.example.com\" data-confirm=\"You can't possibly be sure,\n can you?\">Hello</a>",
-      link_to("Hello", "http://www.example.com", :confirm => "You can't possibly be sure,\n can you?")
+      link_to("Hello", "http://www.example.com", :data => { :confirm => "Are you sure?" })
     )
   end
 
@@ -263,16 +301,35 @@ class UrlHelperTest < ActiveSupport::TestCase
   end
 
   def test_link_tag_using_post_javascript_and_confirm
+    assert_deprecated ":confirm option is deprecated and will be removed from Rails 4.0. Use ':data => { :confirm => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<a href=\"http://www.example.com\" data-method=\"post\" rel=\"nofollow\" data-confirm=\"Are you serious?\">Hello</a>",
+        link_to("Hello", "http://www.example.com", :method => :post, :confirm => "Are you serious?")
+      )
+    end
+  end
+
+  def test_link_tag_using_post_javascript_and_confirm_without_deprecation_warning
     assert_dom_equal(
       "<a href=\"http://www.example.com\" data-method=\"post\" rel=\"nofollow\" data-confirm=\"Are you serious?\">Hello</a>",
-      link_to("Hello", "http://www.example.com", :method => :post, :confirm => "Are you serious?")
+      link_to("Hello", "http://www.example.com", :method => :post, :data => { :confirm => "Are you serious?" })
     )
   end
 
   def test_link_tag_using_delete_javascript_and_href_and_confirm
+    assert_deprecated ":confirm option is deprecated and will be removed from Rails 4.0. Use ':data => { :confirm => \'Text\' }' instead" do
+      assert_dom_equal(
+        "<a href='\#' rel=\"nofollow\" data-confirm=\"Are you serious?\" data-method=\"delete\">Destroy</a>",
+        link_to("Destroy", "http://www.example.com", :method => :delete, :href => '#', :confirm => "Are you serious?"),
+        "When specifying url, form should be generated with it, but not this.href"
+      )
+    end
+  end
+
+  def test_link_tag_using_delete_javascript_and_href_and_confirm_without_deprecation_warning
     assert_dom_equal(
       "<a href='\#' rel=\"nofollow\" data-confirm=\"Are you serious?\" data-method=\"delete\">Destroy</a>",
-      link_to("Destroy", "http://www.example.com", :method => :delete, :href => '#', :confirm => "Are you serious?"),
+      link_to("Destroy", "http://www.example.com", :method => :delete, :href => '#', :data => { :confirm => "Are you serious?" }),
       "When specifying url, form should be generated with it, but not this.href"
     )
   end
